tx_json, wctl, cmd/wctl: move JSON code into wctl pkg
Since the JSON representation of transaction payload is only used in
cmd/wctl, it seems out of place to be put in the wavelet pkg.
